Alt_Pipeline/requirements.txt

certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
dcor==0.6
dtw-python==1.3.0
emd==0.6.2
fonttools==4.45.1
idna==3.6
importlib-resources==6.1.1
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.41.1
matlabengine
matplotlib==3.8.2
neurokit2==0.2.7
numba==0.58.1
numpy==1.26.2
pandas==2.1.3
Pillow==10.1.0
pyparsing==3.1.1
pytz==2023.3.post1
PyWavelets==1.5.0
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
sparse==0.14.0
tabulate==0.9.0
threadpoolctl==3.2.0
tzdata==2023.3
urllib3==2.1.0
zipp==3.17.0
